CA-781 Moved useAppState to its own file
