+++ src/components/MenuBar/ConnectionOptions/ConnectionOptions.tsx
@@ -14,9 +14,9 @@ import {
 import { makeStyles } from '@material-ui/core/styles';
 import { inputLabels, Settings } from '../../../state/settings/settingsReducer';
 import { RenderDimensions } from '../../../state/settings/renderDimensions';
-import { useAppState } from '../../../state';
 import useRoomState from '../../../hooks/useRoomState/useRoomState';
 import useWindowSize from '../../../hooks/useWindowSize/useWindowSize';
+import { useAppState } from '../../../hooks/useAppState/useAppState';
 
 const useStyles = makeStyles({
   formControl: {
