+++ src/components/ParticipantTracks/ParticipantTracks.test.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import ParticipantTracks from './ParticipantTracks';
 import { shallow } from 'enzyme';
 import usePublications from '../../hooks/usePublications/usePublications';
-import { useAppState } from '../../state';
+import { useAppState } from '../../hooks/useAppState/useAppState';
 
-jest.mock('../../state');
+jest.mock('../../hooks/useAppState/useAppState');
 jest.mock('../../hooks/usePublications/usePublications', () =>
   jest.fn(() => [
     { trackSid: 0, kind: 'video', trackName: '' },
