+++ src/components/Room/Room.tsx
@@ -3,7 +3,7 @@ import ParticipantStripCollaboration from '../ParticipantStrip/ParticipantStripC
 import ParticipantStripGrid from '../ParticipantStrip/ParticipantStripGrid/ParticipantStripGrid';
 import { styled } from '@material-ui/core/styles';
 import MainParticipant from '../MainParticipant/MainParticipant';
-import { useAppState } from '../../state';
+import { useAppState } from '../../hooks/useAppState/useAppState';
 
 const Container = styled('div')(({ theme }) => ({
   position: 'relative',
