+++ src/hooks/useAppState/useAppState.tsx
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+import { StateContext } from '../../state';
+
+export function useAppState(): any {
+  const context = useContext(StateContext);
+  if (!context) {
+    throw new Error('useAppState must be used within the AppStateProvider');
+  }
+  return context;
+}
