+++ src/hooks/useParticipant/useParticipant.tsx
@@ -1,8 +1,8 @@
 import useVideoContext from '../useVideoContext/useVideoContext';
 import { TRACK_TYPE } from '../../utils/displayStrings';
 import { Participant } from 'twilio-video';
-import { useAppState } from '../../state';
 import { IMuteRemoteParticipantMessage } from '../../utils/muteRemoteParticipantMessage';
+import { useAppState } from '../useAppState/useAppState';
 
 export interface IParticipantCommands {
   muteParticipant: (participantToMute: Participant) => void;
