+++ src/state/index.tsx
@@ -1,48 +1,13 @@
-import React, { createContext, useContext, useReducer, useState } from 'react';
+import React, { createContext, useReducer, useState } from 'react';
 import { TwilioError } from 'twilio-video';
 import { NOTIFICATION_MESSAGE } from '../utils/displayStrings';
 import axios from 'axios';
 import { ROLE_PERMISSIONS } from '../utils/rbac/rolePermissions';
-import { settingsReducer, initialSettings, Settings, SettingsAction } from './settings/settingsReducer';
+import { settingsReducer, initialSettings } from './settings/settingsReducer';
 import * as jwt_decode from 'jwt-decode';
 import roleChecker from '../utils/rbac/roleChecker';
 import useConfig from '../hooks/useConfig/useConfig';
-
-export interface ParticipantInformation {
-  caseReference: string;
-  displayName: string;
-  partyType: string;
-  userId: number | null;
-  videoConferenceRoomName: string;
-}
-
-export interface StateContextType {
-  error: TwilioError | null;
-  setError(error: TwilioError | null): void;
-  notification: string | null;
-  setNotification(notification: string | null): void;
-  isAutoRetryingToJoinRoom: boolean;
-  disconnectParticipant(isRegistered?: boolean): void;
-  setIsAutoRetryingToJoinRoom(isAutoRetrying: boolean): void;
-  waitingNotification: string;
-  setWaitingNotification(waitingNotification: string | null): void;
-  isFetching: boolean;
-  setSelectedAudioInput: string;
-  selectedVideoInput: string;
-  setSelectedVideoInput: string;
-  selectedSpeakerOutput: string;
-  setSelectedSpeakerOutput: string;
-  gridView: boolean;
-  setGridView: any;
-  authoriseParticipant(): Promise<any>;
-  participantInfo: ParticipantInformation;
-  getToken(participantInformation: ParticipantInformation): Promise<string>;
-  removeParticipant: any;
-  activeSinkId: string;
-  setActiveSinkId(sinkId: string): void;
-  settings: Settings;
-  dispatchSetting: React.Dispatch<SettingsAction>;
-}
+import { ParticipantInformation, StateContextType } from '../types';
 
 export const StateContext = createContext<StateContextType>(null!);
 
@@ -210,12 +175,4 @@ export default function AppStateProvider(props: React.PropsWithChildren<{}>) {
       {props.children}
     </StateContext.Provider>
   );
-}
-
-export function useAppState(): any {
-  const context = useContext(StateContext);
-  if (!context) {
-    throw new Error('useAppState must be used within the AppStateProvider');
-  }
-  return context;
 }
